Fixes multithreaded corruption issues on stacks
This commit moves the filter stack and time filter stack instantiations
from the __init__ method of the module class into the open() method of
the module class.

The reason behind this change is that, had ever mod-logstore receive
multiple Livestatus requests at the same time in a multiprocessor
system, multiple threads will handle these requests concurrently.
Because the threads perform a shallow copy of the logstore instance once
they are in a different thread, the references to the stack will be the
same in every thread, so when different threads process different
filters, they will be added into the same, global, filter stack, which
will cause issues once these requests are converted into SQL because
both queries will be corrupted: one of these will have filters of
multiple requests, and the others will have empty filters (1 = 1).

By moving the stack initialization into open(), we guarantee that each
LivestatusClientThread will have their own copy of the stack, because
the open() method is called after successfully forking the instance,
which is why this method is used to open the database anyway as it
wouldn't be safe to share the same database instance among multiple
threads.